add symbexpr (#204)
* add range array test

* add graphviz for ConstraintGraph

* add Refop

* fix SSA bugs on array

* add aggregate_op

* fix RefOp eval

* fix Ref range

* rust_interval varsion

* add symbolic testcase

* add Symb

* fix varnode bugs

* fix RefOp eval

* rust_interval varsion

* add Symb

* fix varnode bugs

* add print_symbexpr

* fix add_varnode_sym

---------

+++ rapx/src/analysis/senryx/visitor.rs
@@ -608,7 +608,7 @@ impl<'tcx> BodyVisitor<'tcx> {
     }
 
     pub fn get_all_paths(&mut self) -> HashMap<Vec<usize>, Vec<(Place<'tcx>, Place<'tcx>, BinOp)>> {
-        let mut range_analyzer = RangeAnalyzer::<i128>::new(self.tcx, false);
+        let mut range_analyzer = RangeAnalyzer::<i64>::new(self.tcx, false);
         let path_constraints_option =
             range_analyzer.start_path_constraints_analysis_for_defid(self.def_id); // if def_id does not exist, this will break down
         let mut path_constraints: HashMap<Vec<usize>, Vec<(_, _, _)>> =
