+++ rapx/src/lib.rs
@@ -482,19 +482,19 @@ pub fn start_analyzer(tcx: TyCtxt, callback: &RapCallback) {
     if callback.is_range_analysis_enabled() {
         match callback.range {
             1 => {
-                let mut analyzer = RangeAnalyzer::<i128>::new(tcx, false);
+                let mut analyzer = RangeAnalyzer::<i64>::new(tcx, false);
                 analyzer.run();
                 let result = analyzer.get_all_fn_ranges();
                 rap_info!("{}", RAResultMapWrapper(result));
             }
             2 => {
-                let mut analyzer = RangeAnalyzer::<i128>::new(tcx, true);
+                let mut analyzer = RangeAnalyzer::<i64>::new(tcx, true);
                 analyzer.run();
                 let result = analyzer.get_all_fn_ranges();
                 rap_info!("{}", RAResultMapWrapper(result));
             }
             3 => {
-                let mut analyzer = RangeAnalyzer::<i128>::new(tcx, false);
+                let mut analyzer = RangeAnalyzer::<i64>::new(tcx, false);
                 analyzer.start_path_constraints_analysis();
                 let result = analyzer.get_all_path_constraints();
                 rap_info!("{}", PathConstraintMapWrapper(result));
